refactor(wip): organize existing code
