+++ tsconfig.json
@@ -3,7 +3,7 @@
     "target": "ES2020",
     "moduleResolution": "Node",
   },
-  "files": [
-    "src/index.ts"
+  "include": [
+    "src/**/*"
   ],
 }
